Fix sporadic failure in rlike test
See conversation [here](https://github.com/typelevel/frameless/pull/342#issuecomment-448740999).

The `rlike` test was generating random strings and using them as regular
expressions. This occasionally caused issues when the strings did not
form valid regular expressions (unclosed braces, etc).

This PR solves the issue by using the library
[irrec](https://github.com/ceedubs/irrec). irrec provides Scalacheck generators
for regular expressions and candidate matches for them. Full disclosure: I'm the
only maintainer/contributor for irrec, and it should be considered a hobby
project. Having said that, it's only being added as a test dependency and it
fixes a sporadic build failure, so it may be worth it. @imarios suggested that
it would make a good PR
[here](https://github.com/typelevel/frameless/pull/342#issuecomment-449084595).

Feel free to merge if you like this, but I won't be offended if you don't :)